fiddle with printf format strings and arguments thereof to avoid warnings

+++ raidutil/listdev.cpp
@@ -2339,7 +2339,7 @@ Command::Dpt_Error	List_Device::List_Tags(
 		if (more_Devs_Left)
 		{
       	char tempBuf[5];
-		   sprintf (tempBuf, "%d", this_Objs_Tag);
+		   sprintf (tempBuf, "%d", (int)this_Objs_Tag);
    		output.add_Item (tempBuf);
 
   			err |= List_This_Device (output, this_Objs_Tag, devs_Type);
